+++ library/grid_db.py
@@ -9,11 +9,10 @@ from PIL import Image
 from utils.bootstrap import download_file
 
 API_KEY = ""
-# API_URL = "https://www.steamgriddb.com/api/v2/"
+SGDB_URL = "https://www.steamgriddb.com/api/v2/"
 API_URL = "https://api.panyolsoft.com/steamgriddb/"
 
 # To-do list
-# We need to load steamgriddb api if user provided their own key
 # We need to add handling for cases where service is down or there's no internet
 
 class GameIdentifier(object):